+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added keepsake randomization at the start of a new run (disabled by default, can be enabled in the config file).
+
+### Changed
+- Renamed `RandomizeBoons` to `RandomizeBoonOfferings` in the config for clarity.
 
 ## [1.0.2] - 2024-05-18